+++ bundlers/rollup/index.js
@@ -5,18 +5,13 @@ const plugins = require('./plugins');
 function RollupBundler(app) {
     Bundler.call(this, app);
 
-    this.config.variables.push(`const production = !process.env.ROLLUP_WATCH;`);
-    this.app.packager.s('clean', 'rimraf public/build', ['rimraf']);
-    this.app.packager.s('build', 'rollup -c', ['rollup']);
-    this.app.packager.s('dev', 'rollup -c -w', ['rollup']);
-    this.app.packager.s('start', 'sirv public', ['sirv-cli']);
+    this.v('prod', `!process.env.ROLLUP_WATCH`);
 
-    if (this.app.needsPreprocess()) {
-        this.p(plugins.sveltePluginWithPreprocess);
-    }
-    else {
-        this.p(plugins.sveltePlugin);
-    }
+    this.s('clean', 'rimraf public/build', ['rimraf']);
+    this.s('build', 'rollup -c', ['rollup']);
+    this.s('dev', 'rollup -c -w', ['rollup']);
+    this.s('start', 'sirv public', ['sirv-cli']);
+
     this.p(plugins.cssPlugin);
     this.p(plugins.resolvePlugin);
     this.p(plugins.commonjsPlugin);
@@ -24,18 +19,7 @@ function RollupBundler(app) {
     this.p(plugins.livereloadPlugin);
     this.p(plugins.terserPlugin);
 
-    this.transpiler = function (type) {
-        switch (type) {
-            case 'none':
-                this.config.entry = `'src/main.js'`;
-                break;
-            case 'typescript':
-                this.config.entry = `'src/main.ts'`;
-                this.p(plugins.typescriptPlugin);
-                break;
-        }
-    }
+    this.transpiler.init();
 }
-
 
 module.exports = RollupBundler;