+++ bundlers/webpack/plugins.js
@@ -1,82 +1,60 @@
-function svelteRule() {
-    rule(
-        {
-            test: /\.svelte$/,
-            use: {
-                loader: 'svelte-loader',
-                options: {
-                    compilerOptions: {
-                        dev: !prod
-                    },
-                    emitCss: prod,
-                    hotReload: !prod
-                }
-            }
+const svelteRule = () => (
+    {
+        test: /\.svelte$/,
+        use: {
+            loader: 'svelte-loader',
+            options: { options }
         }
-    )
-}
+    }
+)
 
-svelteRule.prototype.requires = ['svelte', 'svelte-loader']
+svelteRule.requires = ['svelte', 'svelte-loader']
 
-function svelteRuleWithPreprocess() {
-    rule(
-        {
-            test: /\.svelte$/,
-            use: {
-                loader: 'svelte-loader',
-                options: {
-                    compilerOptions: {
-                        dev: !prod
-                    },
-                    emitCss: prod,
-                    hotReload: !prod,
-                    preprocess: sveltePreprocess({ sourceMap: !prod })
-                }
-            }
-        }
-    )
-}
-
-svelteRuleWithPreprocess.prototype = { imports: ['svelte-preprocess'], requires: ['svelte', 'svelte-loader'] }
-
-function cssRule() {
-    rule(
-        {
-            test: /\.css$/,
-            use: [
-                MiniCssExtractPlugin.loader,
-                'css-loader'
-            ]
-        }
-    )
-}
-
-cssRule.prototype = { requires: ['css-loader'], requiresPlugin: [miniCssExtractPlugin] }
-
-function typescriptRule() {
-    rule(
-        {
-            test: /\.ts(x)?$/,
-            loader: 'ts-loader',
-            exclude: /node_modules/
-        }
-    )
-}
-
-typescriptRule.prototype.requires = ['ts-loader', '@tsconfig/svelte']
-
-function miniCssExtractPlugin() {
+const miniCssExtractPlugin = () => (
     new MiniCssExtractPlugin({
         filename: '[name].css'
     })
-}
+)
 
-miniCssExtractPlugin.prototype.imports = ['mini-css-extract-plugin']
+miniCssExtractPlugin.imports = ['mini-css-extract-plugin']
+
+const cssRule = () => (
+    {
+        test: /\.css$/,
+        use: [
+            MiniCssExtractPlugin.loader,
+            'css-loader'
+        ]
+    }
+)
+
+cssRule.requires = ['css-loader'];
+cssRule.requiresPlugin = [miniCssExtractPlugin];
+
+const typescriptRule = () => (
+    {
+        test: /\.ts(x)?$/,
+        loader: 'ts-loader',
+        exclude: /node_modules/
+    }
+)
+
+typescriptRule.requires = ['ts-loader', '@tsconfig/svelte']
+
+const coffeeScriptRule = () => (
+    {
+        test: /\.coffee$/,
+        loader: 'coffee-loader',
+        exclude: /node_modules/
+    }
+)
+
+coffeeScriptRule.requires = ['coffeescript', 'coffee-loader'];
 
 module.exports = {
     svelteRule,
-    svelteRuleWithPreprocess,
     cssRule,
     typescriptRule,
+    coffeeScriptRule,
     miniCssExtractPlugin
 }