+++ transpilers/typescript/index.js
@@ -1,24 +1,21 @@
 const path = require('path');
 const Transpiler = require('../Transpiler');
+const Handlebars = require('handlebars');
 
 function TypescriptTranspiler(app) {
     Transpiler.call(this, app);
 
-    app.bundler.transpiler('typescript');
-
     this.modifyFiles = function () {
         app.fh.rn(`/src/main.js`, `/src/main.ts`);
 
-        let appSvelte = app.fh.rpf('/src/App.svelte')
-            .replace(/(?:<script)(( .*?)*?)>/gm, (m, attrs) => `<script${attrs}${!attrs.includes('lang="ts"') ? ' lang="ts"' : ''}>`)
-            .replace(/export let name;/gm, 'export let name: string;');
-
+        let appSvelte = app.fh.rtf('App.svelte.hbs');
+        appSvelte = Handlebars.compile(appSvelte)({ lang: 'ts', ts: true });
         app.fh.wpf(`/src/App.svelte`, appSvelte);
     }
 
     this.exportConfig = function () {
         this.configFile = 'tsconfig.json';
-        let template = app.fh.rf(path.join(__dirname, './config.template'));
+        let template = app.fh.rf(path.join(__dirname, './exports.hbs'));
         /* change config */
 
         app.fh.wf(path.join(app.dir, this.configFile), template);